# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/property-images": {
      "post": {
        "operationId": "PropertyImagesController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Required role. Allowed: admin | superuser",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyImageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Request successful"
                    },
                    "data": {
                      "$ref": "#/components/schemas/PropertyImageResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid role header"
          },
          "403": {
            "description": "Forbidden. Required role(s): admin, superuser"
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Validation failed"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": [
                        "email must be an email",
                        "price must be a positive number"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Create a new property-image",
        "tags": [
          "PropertyImages"
        ]
      },
      "get": {
        "operationId": "PropertyImagesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Request successful"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PropertyImageResponseDto"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "List all property-images",
        "tags": [
          "PropertyImages"
        ]
      }
    },
    "/api/v1/property-images/{id}": {
      "get": {
        "operationId": "PropertyImagesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "PropertyImage ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Request successful"
                    },
                    "data": {
                      "$ref": "#/components/schemas/PropertyImageResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "PropertyImage not found"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2025-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/v1/resource/123"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Get property-image by ID",
        "tags": [
          "PropertyImages"
        ]
      },
      "patch": {
        "operationId": "PropertyImagesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "PropertyImage ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Required role. Allowed: admin | superuser",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePropertyImageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Request successful"
                    },
                    "data": {
                      "$ref": "#/components/schemas/PropertyImageResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid role header"
          },
          "403": {
            "description": "Forbidden. Required role(s): admin, superuser"
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "PropertyImage not found"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2025-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/v1/resource/123"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Validation failed"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": [
                        "email must be an email",
                        "price must be a positive number"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Update a property-image",
        "tags": [
          "PropertyImages"
        ]
      },
      "delete": {
        "operationId": "PropertyImagesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "PropertyImage ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Required role. Allowed: admin | superuser",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Missing or invalid role header"
          },
          "403": {
            "description": "Forbidden. Required role(s): admin, superuser"
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "PropertyImage not found"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2025-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/v1/resource/123"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Delete a property-image",
        "tags": [
          "PropertyImages"
        ]
      }
    },
...
  },
  "info": {
    "title": "Real Estate API",
    "description": "## Real Estate Backend API\n\nThis API powers the real estate platform, supporting property listings, user management, and bookings.\n\n### Authentication & RBAC\nAll protected endpoints require a `role` header indicating the caller's role.\nValid roles: `superuser`, `admin`, `agent`, `seller`, `buyer`\n\n### Response Format\nAll responses follow the envelope: `{ success, statusCode, message, data }`\n\n### Error Format\nAll errors follow: `{ success: false, statusCode, message, errors? }`",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role-header": {
        "type": "apiKey",
        "in": "header",
        "name": "role",
        "description": "User role for RBAC (superuser | admin | agent | seller | buyer)"
      }
    },
    "schemas": {
...
      "PropertyImageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "propertyId": {
            "type": "string",
            "example": "prop_001"
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/img.jpg"
          },
          "caption": {
            "type": "string",
            "example": "Living Room"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "propertyId",
          "imageUrl",
          "caption",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreatePropertyImageDto": {
        "type": "object",
        "properties": {
          "propertyId": {
            "type": "string",
            "example": "prop_001"
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/img.jpg"
          },
          "caption": {
            "type": "string",
            "example": "Living Room"
          }
        },
        "required": [
          "propertyId",
          "imageUrl"
        ]
      },
      "UpdatePropertyImageDto": {
        "type": "object",
        "properties": {
          "propertyId": {
            "type": "string",
            "example": "prop_001"
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/img.jpg"
          },
          "caption": {
            "type": "string",
            "example": "Living Room"
          }
        }
      },
...
    }
  }
}